requirements.txt
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
streamlit==1.11.0
torch==1.11.0
torchvision==0.12.0
matplotlib==3.5.2
matplotlib-inline==0.1.3




